DTP_project/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
tqdm>=4.62.0
requests>=2.26.0 